MULE-11050: decouple code from ApplicationModel that does the macro expansion to a separate class
