+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/dsl/model/ApplicationModel.java
@@ -14,9 +14,7 @@ import static java.lang.Thread.currentThread;
 import static java.util.Arrays.asList;
 import static java.util.Optional.empty;
 import static java.util.Optional.of;
-import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
 import static org.apache.commons.collections.CollectionUtils.disjunction;
 import static org.apache.commons.lang3.StringUtils.isBlank;
 import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
@@ -26,16 +24,8 @@ import static org.mule.runtime.core.exception.Errors.Identifiers.ANY_IDENTIFIER;
 import static org.mule.runtime.dsl.api.xml.DslConstants.CORE_NAMESPACE;
 import static org.mule.runtime.extension.api.util.NameUtils.hyphenize;
 import static org.mule.runtime.extension.api.util.NameUtils.pluralize;
-import org.mule.metadata.api.model.VoidType;
-import org.mule.metadata.api.visitor.MetadataTypeVisitor;
 import org.mule.runtime.api.exception.MuleRuntimeException;
-import org.mule.runtime.api.meta.model.ExtensionModel;
-import org.mule.runtime.api.meta.model.config.ConfigurationModel;
-import org.mule.runtime.api.meta.model.operation.OperationModel;
-import org.mule.runtime.api.meta.model.parameter.ParameterModel;
-import org.mule.runtime.config.spring.dsl.model.extension.xml.GlobalElementComponentModelModelProperty;
-import org.mule.runtime.config.spring.dsl.model.extension.xml.OperationComponentModelModelProperty;
-import org.mule.runtime.config.spring.dsl.model.extension.xml.XmlExtensionModelProperty;
+import org.mule.runtime.config.spring.dsl.model.extension.xml.MacroExpansionModuleModel;
 import org.mule.runtime.config.spring.dsl.processor.ArtifactConfig;
 import org.mule.runtime.config.spring.dsl.processor.ConfigFile;
 import org.mule.runtime.core.api.config.ConfigurationException;
@@ -59,7 +49,6 @@ import java.util.Optional;
 import java.util.Properties;
 import java.util.Set;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 import org.springframework.util.PropertyPlaceholderHelper;
 import org.w3c.dom.Element;
@@ -140,13 +129,7 @@ public class ApplicationModel {
   public static final String PROTOTYPE_OBJECT_ELEMENT = "prototype-object";
   public static final String SINGLETON_OBJECT_ELEMENT = "singleton-object";
   public static final String INTERCEPTOR_STACK_ELEMENT = "interceptor-stack";
-  /**
-   * literal that represents the name of the global element for any given module. If the module's name is math, then
-   * the value of this field will name the global element as <math:config ../>
-   */
-  public static final String MODULE_CONFIG_GLOBAL_ELEMENT_NAME = "config";
 
-  public static final String MODULE_OPERATION_CONFIG_REF = "config-ref";
   public static final ComponentIdentifier ERROR_HANDLER_IDENTIFIER =
       new ComponentIdentifier.Builder().withNamespace(CORE_NAMESPACE).withName(ERROR_HANDLER).build();
   public static final ComponentIdentifier EXCEPTION_STRATEGY_REFERENCE_IDENTIFIER =
@@ -703,240 +686,13 @@ public class ApplicationModel {
   }
 
   /**
-   * Goes through the entire xml mule application looking for the message processors that can be expanded, and then takes
-   * care of the global elements.
+   * We force the current instance of {@link ApplicationModel} to be highly cohesive with {@link MacroExpansionModuleModel}
+   * as it's responsibility of this object to properly initialize and expand every global element/operation into the
+   * concrete set of message processors
    *
    * @param extensionManager extensions that will be used to check if the element has to be expanded.
    */
   private void expandModules(Optional<ExtensionManager> extensionManager) {
-    if (extensionManager.isPresent()) {
-      Map<String, ExtensionModel> extensions = extensionManager.get().getExtensions().stream()
-          .filter(extensionModel -> extensionModel.getModelProperty(XmlExtensionModelProperty.class).isPresent())
-          .collect(toMap(extensionModel -> extensionModel.getName(), identity()));
-
-      if (!extensions.isEmpty()) {
-        createOperationRefEffectiveModel(extensions);
-        createConfigRefEffectiveModel(extensions);
-      }
-    }
+    extensionManager.ifPresent(manager -> new MacroExpansionModuleModel(this, manager.getExtensions()).expand());
   }
-
-  private void createOperationRefEffectiveModel(Map<String, ExtensionModel> extensionManager) {
-    HashMap<Integer, ComponentModel> componentModelsToReplaceByIndex = new HashMap<>();
-
-    executeOnEveryMuleComponentTree(componentModel -> {
-      for (int i = 0; i < componentModel.getInnerComponents().size(); i++) {
-        ComponentModel operationRefModel = componentModel.getInnerComponents().get(i);
-        ComponentIdentifier identifier = operationRefModel.getIdentifier();
-        String identifierName = identifier.getName();
-        if (identifierName.equals(MODULE_CONFIG_GLOBAL_ELEMENT_NAME)) {
-          continue;
-        }
-        ExtensionModel extensionModel = extensionManager.get(identifier.getNamespace());
-        if (extensionModel != null) {
-          //config elements will be worked later on, that's why we are skipping this element
-          OperationModel operationModel = extensionModel.getOperationModel(identifierName)
-              .orElseThrow(() -> new IllegalArgumentException(format("The operation '%s' is missing in the module '%s'",
-                                                                     identifierName, extensionModel.getName())));
-          ComponentModel replacementModel = createOperationInstance(operationRefModel, extensionModel, operationModel);
-          componentModelsToReplaceByIndex.put(i, replacementModel);
-        }
-      }
-      for (Map.Entry<Integer, ComponentModel> entry : componentModelsToReplaceByIndex.entrySet()) {
-        entry.getValue().setParent(componentModel);
-        componentModel.getInnerComponents().add(entry.getKey(), entry.getValue());
-        componentModel.getInnerComponents().remove(entry.getKey() + 1);
-      }
-      componentModelsToReplaceByIndex.clear();
-    });
-  }
-
-  private void createConfigRefEffectiveModel(Map<String, ExtensionModel> extensionManager) {
-    this.muleComponentModels.stream()
-        .forEach(componentModel -> {
-          HashMap<Integer, List<ComponentModel>> componentModelsToReplaceByIndex = new HashMap<>();
-
-          for (int i = 0; i < componentModel.getInnerComponents().size(); i++) {
-            ComponentModel configRefModel = componentModel.getInnerComponents().get(i);
-            ComponentIdentifier identifier = configRefModel.getIdentifier();
-            ExtensionModel extensionModel = extensionManager.get(identifier.getNamespace());
-            if (extensionModel != null) {
-              List<ComponentModel> replacementGlobalElements =
-                  createGlobalElementsInstance(configRefModel, extensionModel);
-              componentModelsToReplaceByIndex.put(i, replacementGlobalElements);
-            }
-          }
-          for (Map.Entry<Integer, List<ComponentModel>> entry : componentModelsToReplaceByIndex.entrySet()) {
-            componentModel.getInnerComponents().addAll(entry.getKey(), entry.getValue());
-            componentModel.getInnerComponents().remove(entry.getKey() + entry.getValue().size());
-          }
-        });
-  }
-
-  private List<ComponentModel> createGlobalElementsInstance(ComponentModel configRefModel, ExtensionModel extensionModel) {
-    List<ComponentModel> globalElementsModel = new ArrayList<>();
-
-    Optional<ConfigurationModel> config = extensionModel.getConfigurationModel("config");
-    if (config.isPresent()
-        && config.get().getModelProperty(GlobalElementComponentModelModelProperty.class).isPresent()) {
-      GlobalElementComponentModelModelProperty globalElementComponentModelModelProperty =
-          config.get().getModelProperty(GlobalElementComponentModelModelProperty.class).get();
-
-      globalElementsModel.addAll(globalElementComponentModelModelProperty.getGlobalElements().stream()
-          .map(globalElementModel -> copyComponentModel(globalElementModel, configRefModel.getNameAttribute()))
-          .collect(Collectors.toList()));
-
-      ComponentModel muleRootElement = configRefModel.getParent();
-      globalElementsModel.stream().forEach(componentModel -> {
-        componentModel.setRoot(true);
-        componentModel.setParameter(NAME_ATTRIBUTE, componentModel.getNameAttribute() + "-" + configRefModel.getNameAttribute());
-        componentModel.setParent(muleRootElement);
-      });
-    }
-    return globalElementsModel;
-  }
-
-  /**
-   * Takes a one liner call to any given message processor, expand it to creating a "module-operation-chain" scope which
-   * has the set of properties, the set of parameters and the list of message processors to execute.
-   *
-   * <pre></pre>
-   * @param operationRefModel message processor that will be replaced by a scope element named "module-operation-chain".
-   * @param extensionModel extension that holds a possible set of <property/>s that has to be parametrized to the new scope.
-   * @param operationModel operation that provides both the <parameter/>s and content of the <body/>
-   * @return a new component model that represents the old placeholder but expanded with the content of the <body/>
-   */
-  private ComponentModel createOperationInstance(ComponentModel operationRefModel, ExtensionModel extensionModel,
-                                                 OperationModel operationModel) {
-    List<ComponentModel> bodyProcessors = operationModel.getModelProperty(OperationComponentModelModelProperty.class).get()
-        .getComponentModel().getInnerComponents();
-
-    String configRefName = operationRefModel.getParameters().get(MODULE_OPERATION_CONFIG_REF);
-
-    ComponentModel.Builder processorChainBuilder = new ComponentModel.Builder();
-    processorChainBuilder
-        .setIdentifier(new ComponentIdentifier.Builder().withNamespace("mule").withName("module-operation-chain").build());
-
-
-    ReturnsVoidTypeVisitor returnsVoidTypeVisitor = new ReturnsVoidTypeVisitor();
-    operationModel.getOutput().getType().accept(returnsVoidTypeVisitor);
-
-    processorChainBuilder.addParameter("returnsVoid", Boolean.toString(returnsVoidTypeVisitor.returnsVoid), false);
-    Map<String, String> propertiesMap = extractProperties(operationRefModel, extensionModel);
-    Map<String, String> parametersMap = extractParameters(operationRefModel, operationModel.getParameterModels());
-    ComponentModel propertiesComponentModel =
-        getParameterChild(propertiesMap, "module-operation-properties", "module-operation-property-entry");
-    ComponentModel parametersComponentModel =
-        getParameterChild(parametersMap, "module-operation-parameters", "module-operation-parameter-entry");
-    processorChainBuilder.addChildComponentModel(propertiesComponentModel);
-    processorChainBuilder.addChildComponentModel(parametersComponentModel);
-
-    for (ComponentModel bodyProcessor : bodyProcessors) {
-      processorChainBuilder.addChildComponentModel(copyComponentModel(bodyProcessor, configRefName));
-    }
-    for (Map.Entry<String, Object> customAttributeEntry : operationRefModel.getCustomAttributes().entrySet()) {
-      processorChainBuilder.addCustomAttribute(customAttributeEntry.getKey(), customAttributeEntry.getValue());
-    }
-    ComponentModel processorChainModel = processorChainBuilder.build();
-    for (ComponentModel processoChainModelChild : processorChainModel.getInnerComponents()) {
-      processoChainModelChild.setParent(processorChainModel);
-    }
-    return processorChainModel;
-  }
-
-  /**
-   * visits all the possible types of a given parameter to realize if it's a "void return type", in which case the
-   * expanded chain will not modify the structure of the event
-   */
-  private class ReturnsVoidTypeVisitor extends MetadataTypeVisitor {
-
-    private boolean returnsVoid = false;
-
-    @Override
-    public void visitVoid(VoidType voidType) {
-      returnsVoid = true;
-    }
-  }
-
-  private ComponentModel getParameterChild(Map<String, String> parameters, String wrapperParameters, String entryParameter) {
-    ComponentModel.Builder parametersBuilder = new ComponentModel.Builder();
-    parametersBuilder
-        .setIdentifier(new ComponentIdentifier.Builder().withNamespace("mule").withName(wrapperParameters).build());
-    parameters.forEach((paramName, paramValue) -> {
-      ComponentModel.Builder parameterBuilder = new ComponentModel.Builder();
-      parameterBuilder.setIdentifier(new ComponentIdentifier.Builder().withNamespace("mule")
-          .withName(entryParameter).build());
-
-      parameterBuilder.addParameter("key", paramName, false);
-      parameterBuilder.addParameter("value", paramValue, false);
-      parametersBuilder.addChildComponentModel(parameterBuilder.build());
-    });
-
-    ComponentModel parametersComponentModel = parametersBuilder.build();
-    for (ComponentModel parameterComponentModel : parametersComponentModel.getInnerComponents()) {
-      parameterComponentModel.setParent(parametersComponentModel);
-    }
-    return parametersComponentModel;
-  }
-
-  private Map<String, String> extractProperties(ComponentModel operationRefModel, ExtensionModel extensionModel) {
-    Map<String, String> valuesMap = new HashMap<>();
-    //extract the <properties>
-    String configParameter = operationRefModel.getParameters().get(MODULE_OPERATION_CONFIG_REF);
-    if (configParameter != null) {
-      ComponentModel configRefComponentModel = this.muleComponentModels.get(0).getInnerComponents().stream()
-          .filter(componentModel -> componentModel.getIdentifier().getNamespace().equals(extensionModel.getName())
-              && componentModel.getIdentifier().getName().equals(MODULE_CONFIG_GLOBAL_ELEMENT_NAME)
-              && configParameter.equals(componentModel.getParameters().get(NAME_ATTRIBUTE)))
-          .findFirst()
-          .orElseThrow(() -> new IllegalArgumentException(String
-              .format("There's no <%s:config> named [%s] in the current mule app", extensionModel.getName(), configParameter)));
-      valuesMap
-          .putAll(extractParameters(configRefComponentModel,
-                                    extensionModel.getConfigurationModel("config").get().getParameterModels()));
-    }
-    return valuesMap;
-  }
-
-  private Map<String, String> extractParameters(ComponentModel componentModel, List<ParameterModel> parameters) {
-    Map<String, String> valuesMap = new HashMap<>();
-    for (ParameterModel parameterExtension : parameters) {
-      String paramName = parameterExtension.getName();
-      String value = null;
-      if (componentModel.getParameters().containsKey(paramName)) {
-        value = componentModel.getParameters().get(paramName);
-      } else if (parameterExtension.getDefaultValue() != null) {
-        value = (String) parameterExtension.getDefaultValue();
-      }
-      valuesMap.put(paramName, value);
-    }
-    return valuesMap;
-  }
-
-  private ComponentModel copyComponentModel(ComponentModel modelToCopy,
-                                            String configRefName) {
-    ComponentModel.Builder operationReplacementModel = new ComponentModel.Builder();
-    operationReplacementModel
-        .setIdentifier(modelToCopy.getIdentifier())
-        .setTextContent(modelToCopy.getTextContent());
-    for (Map.Entry<String, Object> entry : modelToCopy.getCustomAttributes().entrySet()) {
-      operationReplacementModel.addCustomAttribute(entry.getKey(), entry.getValue());
-    }
-    for (Map.Entry<String, String> entry : modelToCopy.getParameters().entrySet()) {
-      String value = entry.getKey().endsWith(REFERENCE_ATTRIBUTE)
-          ? entry.getValue().concat("-").concat(configRefName)
-          : entry.getValue();
-      operationReplacementModel.addParameter(entry.getKey(), value, false);
-    }
-    for (ComponentModel operationChildModel : modelToCopy.getInnerComponents()) {
-      operationReplacementModel.addChildComponentModel(copyComponentModel(//parameterValues,
-                                                                          operationChildModel, configRefName));
-    }
-    ComponentModel componentModel = operationReplacementModel.build();
-    for (ComponentModel child : componentModel.getInnerComponents()) {
-      child.setParent(componentModel);
-    }
-    return componentModel;
-  }
-
 }
